requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
fsspec==2023.9.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
Pillow==10.0.1
pydantic==2.4.2
pydantic_core==2.10.1
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
